refactor(conversations): relocate subtree to features/conversations (Phase 1a)
Mechanical move of `app/src/pages/conversations/{components,hooks,utils,
composerSendDecision,taskPlanActions}` to `app/src/features/conversations/`,
matching the `features/human/` layout per the timeline-refactor plan. Pure
relocation — no behavior change. `pages/` and `features/` sit at the same depth
under `app/src`, so the moved files' own imports are unchanged; only references
*to* the subtree were repointed (Conversations.tsx, pages/dev preview, and 9
external importers under components/{chat,intelligence,flows,notifications,
channels,layout}).

Typecheck + lint clean; 337 tests across the moved subtree and all consumers
(Conversations render suite, HumanPage, Accounts, IntelligenceTasksTab) pass.

Refs: docs/plans/conversations-timeline-refactor.md Phase 1
