+++ app/src/components/flows/WorkflowCopilotPanel.tsx
@@ -25,8 +25,8 @@ import { useWorkflowBuilderChat } from '../../hooks/useWorkflowBuilderChat';
 import { diffGraphs } from '../../lib/flows/graphDiff';
 import type { WorkflowGraph } from '../../lib/flows/types';
 import { useT } from '../../lib/i18n/I18nContext';
-import { BubbleMarkdown } from '../../pages/conversations/components/AgentMessageBubble';
-import { ToolTimelineBlock } from '../../pages/conversations/components/ToolTimelineBlock';
+import { BubbleMarkdown } from '../../features/conversations/components/AgentMessageBubble';
+import { ToolTimelineBlock } from '../../features/conversations/components/ToolTimelineBlock';
 import type { WorkflowProposal } from '../../store/chatRuntimeSlice';
 import ChatComposer from '../chat/ChatComposer';
 import Button from '../ui/Button';
