+++ app/src/features/conversations/components/AgentMessageBubble.tsx
