+++ app/src/features/conversations/components/AgentTimelineRail.tsx
