+++ app/src/features/conversations/components/ThreadGoalChip.tsx
