+++ app/src/features/conversations/components/ThreadTodoStrip.tsx
