+++ app/src/features/conversations/components/ToolFailureLines.tsx
