+++ app/src/features/conversations/components/__tests__/AgentProcessSourcePanel.test.tsx
