+++ app/src/features/conversations/components/__tests__/AgentTimelineRail.test.tsx
