+++ app/src/features/conversations/components/__tests__/BackgroundActivityRows.test.tsx
