+++ app/src/features/conversations/components/__tests__/BackgroundProcessesPanel.test.tsx
