+++ app/src/features/conversations/hooks/useBackgroundActivity.ts
