+++ app/src/pages/dev/AgentInsightsPreview.tsx
@@ -2,8 +2,8 @@ import { useState } from 'react';
 
 import Button from '../../components/ui/Button';
 import type { ToolTimelineEntry } from '../../store/chatRuntimeSlice';
-import { AgentProcessSourcePanel } from '../conversations/components/AgentProcessSourcePanel';
-import { ToolTimelineBlock } from '../conversations/components/ToolTimelineBlock';
+import { AgentProcessSourcePanel } from '../../features/conversations/components/AgentProcessSourcePanel';
+import { ToolTimelineBlock } from '../../features/conversations/components/ToolTimelineBlock';
 
 /**
  * Dev-only visual preview of the "Agentic task insights" Chat surface.
